requirements.txt
openai>=1.0.0
faiss-cpu
numpy
pandas
tqdm
python-dotenv
fastapi
uvicorn
rich
pydantic
PyPDF2>=3.0.0
pdfplumber>=0.9.0
python-magic-bin>=0.4.14
qdrant-client>=1.6.0
